refactor : poll 투표 로직 분리
